refactor(chat-server): extract file uploads into chat-server/files.ts
Moves the multipart/chunked file upload routes and the file-serve route
out of chat-server.ts's handleHttp into dedicated handlers. Each
handler is a self-contained function that takes (req, res, ...) and
manages its own response.

- chat-server/files.ts (new, 383 lines):
    * Constants: MAX_UPLOAD_SIZE, CHUNK_UPLOAD_TIMEOUT, MIME table
    * State: pendingChunkedUploads Map (per-server)
    * Helpers: cleanupChunkedUpload, resolveGroupFolder, getUploadsDir,
      local json/readBody for self-contained dispatch
    * Exports:
        - handleMultipartUpload(req, res, roomId, identity)
        - handleChunkedUpload(req, res, roomId, identity)  [async]
        - handleFileServe(res, groupFolder, filename)
- chat-server.ts: -312 lines net. handleHttp's three file-route blocks
  collapse to three-line dispatches. Orphan imports removed: Busboy,
  randomUUID, isValidGroupFolder.

chat-server.ts: 1104 -> 792 lines.
