+++ src/chat-server.ts
@@ -24,8 +24,6 @@ import os from 'os';
 import path from 'path';
 import fs from 'fs';
 import { execFile } from 'child_process';
-import Busboy from 'busboy';
-import { randomUUID } from 'crypto';
 
 import { DATA_DIR } from './config.js';
 import { redactSensitiveData } from './redact.js';
@@ -43,7 +41,7 @@ import {
   authenticateRequest,
   warnIfAutoProxyTrust,
 } from './chat-server/auth.js';
-import { isValidGroupFolder, resolveGroupFolderPath } from './group-folder.js';
+import { resolveGroupFolderPath } from './group-folder.js';
 import { logger } from './logger.js';
 import {
   initChatDatabase,
@@ -80,52 +78,14 @@ import {
 } from './chat-server/state.js';
 import { initWebPush, isValidPushEndpoint } from './chat-server/push.js';
 import { setupWebSocket } from './chat-server/ws.js';
+import {
+  handleMultipartUpload,
+  handleChunkedUpload,
+  handleFileServe,
+} from './chat-server/files.js';
 
 export function isChatServerRunning(): boolean {
   return server !== null;
-}
-
-// ── File upload helpers ────────────────────────────────────────────────────
-const MAX_UPLOAD_SIZE = 1024 * 1024 * 1024; // 1GB
-const CHUNK_SIZE = 512 * 1024; // 512KB chunks (stays well under Azure's 2MB limit)
-const CHUNK_UPLOAD_TIMEOUT = 5 * 60 * 1000; // 5 minutes to complete a chunked upload
-
-// Track in-progress chunked uploads
-const pendingChunkedUploads = new Map<
-  string,
-  {
-    groupFolder: string;
-    roomId: string;
-    filename: string;
-    mime: string;
-    totalChunks: number;
-    receivedChunks: Set<number>;
-    tempDir: string;
-    sender: string;
-    timer: ReturnType<typeof setTimeout>;
-    cumulativeSize: number;
-  }
->();
-
-function cleanupChunkedUpload(uploadId: string): void {
-  const upload = pendingChunkedUploads.get(uploadId);
-  if (!upload) return;
-  clearTimeout(upload.timer);
-  try {
-    fs.rmSync(upload.tempDir, { recursive: true, force: true });
-  } catch {}
-  pendingChunkedUploads.delete(uploadId);
-}
-
-function resolveGroupFolder(roomId: string): string | null {
-  const jid = `chat:${roomId}`;
-  const groups = getAllRegisteredGroups();
-  const group = groups[jid];
-  return group ? group.folder : null;
-}
-
-function getUploadsDir(groupFolder: string): string {
-  return path.join(resolveGroupFolderPath(groupFolder), 'uploads');
 }
 
 // ── Static PWA serving ─────────────────────────────────────────────────────
@@ -661,106 +621,10 @@ async function handleHttp(
     return;
   }
 
-  // ── File upload ─────────────────────────────────────────────────────────
+  // ── File upload (multipart) ─────────────────────────────────────────────
   const uploadMatch = url.pathname.match(/^\/api\/rooms\/([^/]+)\/upload$/);
   if (uploadMatch && method === 'POST') {
-    const roomId = uploadMatch[1];
-    const groupFolder = resolveGroupFolder(roomId);
-    if (!groupFolder)
-      return json(404, { error: 'Room not registered as a group' });
-
-    const uploadsDir = getUploadsDir(groupFolder);
-    fs.mkdirSync(uploadsDir, { recursive: true });
-
-    const contentType = req.headers['content-type'] || '';
-    if (!contentType.includes('multipart/form-data')) {
-      return json(400, { error: 'Content-Type must be multipart/form-data' });
-    }
-
-    const busboy = Busboy({
-      headers: req.headers,
-      limits: { fileSize: MAX_UPLOAD_SIZE, files: 1 },
-    });
-    let fileInfo: {
-      id: string;
-      filename: string;
-      mime: string;
-      size: number;
-      path: string;
-    } | null = null;
-    let limitHit = false;
-    let caption = '';
-
-    busboy.on('field', (name, value) => {
-      if (name === 'caption') caption = value.trim();
-    });
-
-    busboy.on('file', (_fieldname, stream, info) => {
-      const id = randomUUID();
-      const ext = path.extname(info.filename) || '';
-      const safeFilename = `${id}${ext}`;
-      const filePath = path.join(uploadsDir, safeFilename);
-      let size = 0;
-
-      const ws = fs.createWriteStream(filePath);
-      stream.on('data', (chunk: Buffer) => {
-        size += chunk.length;
-      });
-      stream.pipe(ws);
-
-      stream.on('limit', () => {
-        limitHit = true;
-        ws.destroy();
-        try {
-          fs.unlinkSync(filePath);
-        } catch {}
-      });
-
-      stream.on('end', () => {
-        if (!limitHit) {
-          fileInfo = {
-            id,
-            filename: info.filename,
-            mime: info.mimeType || 'application/octet-stream',
-            size,
-            path: `/api/files/${encodeURIComponent(groupFolder)}/${safeFilename}`,
-          };
-        }
-      });
-    });
-
-    busboy.on('finish', async () => {
-      if (limitHit)
-        return json(413, {
-          error: `File exceeds ${(MAX_UPLOAD_SIZE / 1024 / 1024 / 1024).toFixed(1)}GB limit`,
-        });
-      if (!fileInfo) return json(400, { error: 'No file uploaded' });
-
-      const sender = senderIdentity;
-      const fileMeta = {
-        url: fileInfo.path,
-        filename: fileInfo.filename,
-        mime: fileInfo.mime,
-        size: fileInfo.size,
-      };
-      const stored = storeFileMessage(
-        roomId,
-        sender,
-        'user',
-        fileMeta,
-        caption,
-      );
-      broadcast(roomId, { type: 'message', ...stored });
-
-      // Trigger the channel adapter so the agent processes this message
-      getOnNewMessage()?.(roomId, stored);
-
-      json(200, { ...fileInfo, caption });
-    });
-
-    busboy.on('error', () => json(500, { error: 'Upload failed' }));
-    req.pipe(busboy);
-    return;
+    return handleMultipartUpload(req, res, uploadMatch[1], senderIdentity);
   }
 
   // ── Web Push: VAPID public key + subscribe/unsubscribe ──────────────────
@@ -836,197 +700,22 @@ async function handleHttp(
     return json(200, { ok: true });
   }
 
-  // ── Chunked file upload ─────────────────────────────────────────────────
+  // ── File upload (chunked) ───────────────────────────────────────────────
   const chunkMatch = url.pathname.match(
     /^\/api\/rooms\/([^/]+)\/upload\/chunk$/,
   );
   if (chunkMatch && method === 'POST') {
-    const roomId = chunkMatch[1];
-    const body = await readBody();
-    let parsed: {
-      uploadId: string;
-      chunkIndex: number;
-      totalChunks: number;
-      filename: string;
-      mime: string;
-      data: string;
-      caption?: string;
-    };
-    try {
-      parsed = JSON.parse(body);
-    } catch {
-      return json(400, { error: 'Invalid JSON' });
-    }
-
-    const { uploadId, chunkIndex, totalChunks, filename, mime, data } = parsed;
-    if (!uploadId || chunkIndex == null || !totalChunks || !filename || !data) {
-      return json(400, { error: 'Missing required fields' });
-    }
-
-    // Validate uploadId as UUID to prevent path traversal
-    if (
-      !/^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i.test(
-        uploadId,
-      )
-    ) {
-      return json(400, { error: 'Invalid uploadId format' });
-    }
-
-    const groupFolder = resolveGroupFolder(roomId);
-    if (!groupFolder)
-      return json(404, { error: 'Room not registered as a group' });
-
-    // Initialize or retrieve upload state
-    let upload = pendingChunkedUploads.get(uploadId);
-    if (!upload) {
-      const tempDir = path.join(os.tmpdir(), `nanoclaw-chunk-${uploadId}`);
-      fs.mkdirSync(tempDir, { recursive: true });
-      upload = {
-        groupFolder,
-        roomId,
-        filename,
-        mime: mime || 'application/octet-stream',
-        totalChunks,
-        receivedChunks: new Set(),
-        tempDir,
-        sender: senderIdentity,
-        timer: setTimeout(
-          () => cleanupChunkedUpload(uploadId),
-          CHUNK_UPLOAD_TIMEOUT,
-        ),
-        cumulativeSize: 0,
-      };
-      pendingChunkedUploads.set(uploadId, upload);
-    } else {
-      // Validate totalChunks matches initial value
-      if (totalChunks !== upload.totalChunks) {
-        return json(400, { error: 'totalChunks mismatch' });
-      }
-    }
-
-    // Write chunk and check incremental size
-    const chunkBuf = Buffer.from(data, 'base64');
-    upload.cumulativeSize += chunkBuf.length;
-    if (upload.cumulativeSize > MAX_UPLOAD_SIZE) {
-      cleanupChunkedUpload(uploadId);
-      return json(413, {
-        error: `File exceeds ${(MAX_UPLOAD_SIZE / 1024 / 1024 / 1024).toFixed(1)}GB limit`,
-      });
-    }
-    fs.writeFileSync(path.join(upload.tempDir, String(chunkIndex)), chunkBuf);
-    upload.receivedChunks.add(chunkIndex);
-
-    // Check if all chunks received
-    if (upload.receivedChunks.size < upload.totalChunks) {
-      return json(200, {
-        ok: true,
-        received: upload.receivedChunks.size,
-        total: upload.totalChunks,
-      });
-    }
-
-    // All chunks received — reassemble
-    clearTimeout(upload.timer);
-
-    // Check total size BEFORE writing to prevent disk DoS
-    let totalSize = 0;
-    for (let i = 0; i < totalChunks; i++) {
-      const chunkPath = path.join(upload.tempDir, String(i));
-      try {
-        totalSize += fs.statSync(chunkPath).size;
-      } catch {
-        // Missing chunk
-      }
-    }
-
-    if (totalSize > MAX_UPLOAD_SIZE) {
-      fs.rmSync(upload.tempDir, { recursive: true, force: true });
-      pendingChunkedUploads.delete(uploadId);
-      return json(413, {
-        error: `File exceeds ${(MAX_UPLOAD_SIZE / 1024 / 1024 / 1024).toFixed(1)}GB limit`,
-      });
-    }
-
-    // Write reassembled file and await completion
-    const uploadsDir = getUploadsDir(groupFolder);
-    fs.mkdirSync(uploadsDir, { recursive: true });
-    const id = randomUUID();
-    const ext = path.extname(filename) || '';
-    const safeFilename = `${id}${ext}`;
-    const finalPath = path.join(uploadsDir, safeFilename);
-
-    const writeStream = fs.createWriteStream(finalPath);
-    for (let i = 0; i < totalChunks; i++) {
-      const chunkPath = path.join(upload.tempDir, String(i));
-      writeStream.write(fs.readFileSync(chunkPath));
-    }
-    await new Promise<void>((resolve, reject) => {
-      writeStream.on('finish', resolve);
-      writeStream.on('error', reject);
-      writeStream.end();
-    });
-
-    // Cleanup temp
-    fs.rmSync(upload.tempDir, { recursive: true, force: true });
-    pendingChunkedUploads.delete(uploadId);
-
-    const fileMeta = {
-      url: `/api/files/${encodeURIComponent(groupFolder)}/${safeFilename}`,
-      filename,
-      mime: upload.mime,
-      size: totalSize,
-    };
-    const caption = parsed.caption || '';
-    const stored = storeFileMessage(
-      roomId,
-      upload.sender,
-      'user',
-      fileMeta,
-      caption,
-    );
-    broadcast(roomId, { type: 'message', ...stored });
-    getOnNewMessage()?.(roomId, stored);
-
-    return json(200, { ...fileMeta, caption });
+    return handleChunkedUpload(req, res, chunkMatch[1], senderIdentity);
   }
 
   // Serve uploaded files
   const fileMatch = url.pathname.match(/^\/api\/files\/([^/]+)\/([^/]+)$/);
   if (fileMatch && method === 'GET') {
-    const groupFolder = decodeURIComponent(fileMatch[1]);
-    const filename = decodeURIComponent(fileMatch[2]);
-    // Prevent path traversal
-    if (
-      !isValidGroupFolder(groupFolder) ||
-      filename.includes('..') ||
-      filename.includes('/')
-    ) {
-      res.writeHead(403);
-      res.end();
-      return;
-    }
-    const filePath = path.join(getUploadsDir(groupFolder), filename);
-    const ext = path.extname(filename);
-    const mime = MIME[ext] || 'application/octet-stream';
-    let stat: fs.Stats;
-    try {
-      stat = fs.statSync(filePath);
-    } catch {
-      return json(404, { error: 'File not found' });
-    }
-    res.writeHead(200, {
-      'Content-Type': mime,
-      'Content-Length': stat.size,
-      'Content-Disposition': `inline; filename="${filename}"`,
-      'Cache-Control': 'public, max-age=31536000, immutable',
-      // Sandbox the response into an opaque origin so HTML/SVG uploads
-      // cannot read the PWA's localStorage token. nosniff stops the
-      // browser from reinterpreting the Content-Type.
-      'Content-Security-Policy': 'sandbox',
-      'X-Content-Type-Options': 'nosniff',
-    });
-    fs.createReadStream(filePath).pipe(res);
-    return;
+    return handleFileServe(
+      res,
+      decodeURIComponent(fileMatch[1]),
+      decodeURIComponent(fileMatch[2]),
+    );
   }
 
   // OPTIONS preflight
@@ -1044,7 +733,6 @@ async function handleHttp(
 
   json(404, { error: 'Not found' });
 }
-
 
 // ── Public API ─────────────────────────────────────────────────────────────
 let server: http.Server | https.Server | null = null;
